backend/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Chatbot Tool Calling"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "alembic>=1.18.4",
    "pydantic-settings>=2.13.0",
    "sqlalchemy>=2.0.46",
    "langchain>=1.2.10",
    "langchain-core>=1.2.14",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "psycopg[binary]>=3.3.3",
    "uuid7>=0.1.0",
    "pgvector>=0.3.6",
    "pypdf>=5.0.0",
    "python-multipart>=0.0.18",
    "langchain-text-splitters>=0.3.0",
]
